chore: Move tests into unit, integration, and API folders
